WIP trying to glue all new concepts together - also added a validator for the JSON data coming in and created a unit test for the validator that cannot be started at the moment as the script is too scattered by now, but I am pretty much certain that the concept should be working properly

+++ Scripts/ViewModel/ComponentDraftMapper/DTOs/Connection.cs
@@ -4,7 +4,7 @@
     {
         public int fromPinNr { get; set; }
         public int toPinNr { get; set; }
-        public int magnitude { get; set; }
-        public int phaseInDegrees { get; set; }
+        public float magnitude { get; set; }
+        public float phaseInDegrees { get; set; }
     }
 }
